% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepvec.R
\name{grep_any}
\alias{grep_any}
\alias{\%grepin\%}
\alias{\%igrepin\%}
\title{Check if a string contains any of a vector of patterns.}
\usage{
grep_any(
  haystacks,
  needles,
  ignore_case = FALSE,
  fixed = FALSE,
  use_bytes = FALSE,
  invert = FALSE
)

x \%grepin\% patterns

x \%igrepin\% patterns
}
\description{
Check if a string contains any of a vector of patterns.
}
