% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepvec.R
\name{grep_count}
\alias{grep_count}
\title{Count the number of matches in a vector of patterns for each string.}
\usage{
grep_count(
  haystacks,
  needles,
  ignore_case = FALSE,
  fixed = FALSE,
  use_bytes = FALSE,
  invert = FALSE
)
}
\description{
Count the number of matches in a vector of patterns for each string.
}
