% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepvec.R
\name{grep_first}
\alias{grep_first}
\title{Get the first match from a vector of patterns for each string.}
\usage{
grep_first(
  haystacks,
  needles,
  ignore_case = FALSE,
  value = FALSE,
  fixed = FALSE,
  use_bytes = FALSE,
  invert = FALSE
)
}
\description{
Get the first match from a vector of patterns for each string.
}
