% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepvec.R
\name{grepl_any}
\alias{grepl_any}
\alias{\%grepin\%}
\alias{\%igrepin\%}
\title{Check if a string contains any of a vector of patterns.}
\usage{
grepl_any(
  needles,
  haystacks,
  ignore_case = FALSE,
  fixed = FALSE,
  use_bytes = FALSE,
  invert = FALSE
)

patterns \%grepin\% x

patterns \%igrepin\% x
}
\description{
Check if a string contains any of a vector of patterns.
}
