% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepvec-package.R
\docType{package}
\name{grepvec-package}
\alias{grepvec-package}
\alias{_PACKAGE}
\title{grepvec package}
\description{
Use \code{grepvec} to search a vector of strings for matches in a vector
of regular expressions or fixed strings. R's native functions, like grep,
search for a single pattern in a vector of strings. To search for many
possible patterns across a string, some form of looping is required.
\code{grepvec} implements this in C so that it is faster than native R solutions.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/hans-elliott99/grepvec/issues}
}

}
\author{
\strong{Maintainer}: Hans Elliott \email{hanselliott61@gmail.com}

Other contributors:
\itemize{
  \item Ville Laurikari (Author of the TRE library) [contributor, copyright holder]
}

}
\keyword{internal}
