% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepvectors.R
\name{to_hay}
\alias{to_hay}
\title{Get the list of haystack strings that matched to each needle pattern.}
\usage{
to_hay(results, haystacks)
}
\arguments{
\item{results}{List returned from grepvec, where each list index corresponds
to a haystack string and each list element is an integer vector with the
indices of the needle patterns that matched the haystack string.}

\item{haystacks}{Character vector containing the haystack strings that were
searched over by grepvec.}
}
\value{
List of length(needles) of character vectors. Each list index
corresponds to the needle pattern at that index, and the list elements are
the strings from haystack which matched to the given needle pattern.
}
\description{
Converts grepvec results from a list of indices, length(haystacks), into a
list of haystack strings, length(needles).
}
