% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepvectors.R
\name{to_ndl}
\alias{to_ndl}
\title{Get the list of needle patterns that matched to each haystack string.}
\usage{
to_ndl(results, needles)
}
\arguments{
\item{results}{List returned from grepvec, where each list index corresponds
to a haystack string and each list element is an integer vector with the
indices of the needle patterns that matched the haystack string.}

\item{needles}{Character vector containing the needle patterns that were
searched for in haystacks by grepvec.}
}
\value{
List of length(haystacks) of character vectors. Each list index
corresponds to the haystack string at that index, and the list elements are
the patterns from needles which matched to the given haystack string.
}
\description{
Converts grepvec results from the list of indices into a list of the needle
patterns, length(haystacks).
}
