% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepvec-pkg.R
\docType{package}
\name{grepvec-package}
\alias{grepvec-package}
\alias{_PACKAGE}
\title{grepvec package}
\description{
Use \code{grepvec} to search a vector of strings for matches in a vector
of regular expressions or fixed strings. R's native functions, like grep,
search for a single pattern in a vector of strings. To search for many
possible patterns across a string, some form of looping is required.
\code{grepvec} implements this in C so that it is faster than native R solutions.
}
\author{
\strong{Maintainer}: Hans Elliott \email{hanselliott61@gmail.com}

}
\keyword{internal}
